pyproject.toml
[project]
name = "malicious-url-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autogluon-multimodal>=1.3.1",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch>=2.6.0",
    "transformers>=4.49.0",
]
